overlay: refactor configuration parsing
Problem: overlay.c is huge and needs some refactoring for clarity.

Move configuration processing to ovconf.[ch].
